Do not clear migration table

+++ swagger_server/test/__init__.py
@@ -14,15 +14,13 @@ from swagger_server.encoder import JSONEncoder
 
 import project.app
 
-from user_data_store import models
-
 
 DB = SQLAlchemy()
+
 
 class BaseTestCase(TestCase):
 
     def create_app(self):
-        #logging.getLogger('connexion.operation').setLevel('ERROR')
         app = connexion.App(__name__, specification_dir='../swagger/')
         flask_app = app.app
         flask_app.json_encoder = JSONEncoder
@@ -46,6 +44,9 @@ class BaseTestCase(TestCase):
 
         # By reversing the tables, children should get deleted before parents.
         for table in reversed(meta.sorted_tables):
+            if table.name == "alembic_version":  # Do not delete migration data
+                continue
+
             DB.session.execute(table.delete())
         DB.session.commit()
